requirements.txt
flash-linear-attention
transformers>=4.30.0
einops>=0.6.0
packaging>=20.0
